+++ src/main/java/org/jabref/gui/maintable/ColumnPreferences.java
@@ -8,11 +8,15 @@ public class ColumnPreferences {
     public static final double ICON_COLUMN_WIDTH = 16 + 12; // add some additional space to improve appearance
 
     private final List<MainTableColumnModel> columns;
+
     private final List<MainTableColumnModel> columnSortOrder;
 
-    public ColumnPreferences(List<MainTableColumnModel> columns, List<MainTableColumnModel> columnSortOrder) {
+    private final boolean dedicatedFileColumnsEnabled;
+
+    public ColumnPreferences(List<MainTableColumnModel> columns, List<MainTableColumnModel> columnSortOrder, boolean dedicatedFileColumnsEnabled) {
         this.columns = columns;
         this.columnSortOrder = columnSortOrder;
+        this.dedicatedFileColumnsEnabled = dedicatedFileColumnsEnabled;
     }
 
     public List<MainTableColumnModel> getColumns() {
@@ -22,4 +26,9 @@ public class ColumnPreferences {
     public List<MainTableColumnModel> getColumnSortOrder() {
         return columnSortOrder;
     }
+
+    public boolean isDedicatedFileColumnsEnabled() {
+        return dedicatedFileColumnsEnabled;
+    }
+
 }